requirements.txt
# ── Core ML stack ─────────────────────────────────────────────────────────────
torch>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# ── ALS backend (fast C++ implementation) ─────────────────────────────────────
implicit>=0.7.2

# ── Vector database ───────────────────────────────────────────────────────────
faiss-cpu>=1.7.4

# ── API framework ─────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# ── Data processing ───────────────────────────────────────────────────────────
pandas>=2.0.0
pyarrow>=14.0.0        # for .parquet support used by preprocessing

# ── Progress bars ─────────────────────────────────────────────────────────────
tqdm>=4.66.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
httpx>=0.27.0          # FastAPI TestClient dependency

# ── Optional: LLM reranker ────────────────────────────────────────────────────
# Uncomment the provider you need:
# openai>=1.0.0
# anthropic>=0.25.0

# ── Optional: notebooks & visualisation ──────────────────────────────────────
jupyter>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
umap-learn>=0.5.6

# ── Optional: ablation table formatting ──────────────────────────────────────
tabulate>=0.9.0
